requirements.txt
langchain==0.2.16
langchain-groq==0.1.9
requests
arxiv
streamlit
python-dotenv
tenacity
reportlab